arujisama_flask/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
PyJWT==1.4.2
PyMySQL==0.9.3
pyparsing==2.4.1.1
pytest==5.0.1
six==1.12.0
SQLAlchemy==1.3.6
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.5.2
